fix: move theme provider to config provider and delete genui-chat wrap
